requirements.txt
fastapi
uvicorn
jinja2
python-multipart
numpy
pillow
opencv-python-headless
tensorflow-macos
tensorflow-metal
